finish favorite handler
