+++ cmd/api/handlers/favorite.go
@@ -13,9 +13,9 @@ import (
 
 func FavoriteAction(c *gin.Context) {
 	type formParam struct {
-		Token string `form:"token" binding:"required"`
-		VideoId int64 `form:"video_id" binding:"required"`
-		ActionType int32 `form:"action_type" binding:"required"`
+		Token      string `form:"token" binding:"required"`
+		VideoId    int64  `form:"video_id" binding:"required"`
+		ActionType int32  `form:"action_type" binding:"required"`
 	}
 
 	var param formParam
@@ -28,15 +28,47 @@ func FavoriteAction(c *gin.Context) {
 	userId := int64(claims[constants.IdentityKey].(float64))
 
 	if param.ActionType == 1 {
-		rpc.CreateFavorite(context.Background(), &favoriterpc.CreateFavoriteRequest{
-			UserId: userId,
-	
+		err := rpc.CreateFavorite(context.Background(), &favoriterpc.CreateFavoriteRequest{
+			UserId:  userId,
+			VideoId: param.VideoId,
 		})
+		if err != nil {
+			SendStatusResponse(c, errno.ConvertErr(err))
+			return
+		}
+	} else if param.ActionType == 2 {
+		err := rpc.DeleteFavorite(c, &favoriterpc.DeleteFavoriteRequest{
+			UserId:  userId,
+			VideoId: param.VideoId,
+		})
+		if err != nil {
+			SendStatusResponse(c, errno.ConvertErr(err))
+		}
+	}
+	SendStatusResponse(c, errno.Success)
+}
 
+func FavoriteList(c *gin.Context) {
+	type formParam struct {
+		UserId int64  `form:"user_id" binding:"required"`
+		Token  string `form:"token" binding:"required"`
 	}
 
+	var param formParam
+	if err := c.ShouldBind(&param); err != nil {
+		SendVideoListResponse(c, err, nil)
+		return
+	}
 
+	claims := jwt.ExtractClaims(c)
+	userId := int64(claims[constants.IdentityKey].(float64))
 
-
-
+	videos, err := rpc.QueryFavorite(context.Background(), &favoriterpc.QueryFavoriteByUserIdRequest{
+		UserId: userId,
+	})
+	if err != nil {
+		SendVideoListResponse(c, err, nil)
+		return
+	}
+	SendVideoListResponse(c, errno.Success, videos)
 }
