--- idl/core/favorite.proto
@@ -1,54 +1,0 @@
-syntax = "proto3";
-package douyin.core;
-
-option go_package = "douyin/core";
-// /douyin/favorite - 点赞接口
-// 接口类型 POST
-
-message ActionRequest{
-    int64 user_id=1;//用户id
-    string token=2;//用户鉴权token
-    int64 video_id=3;//视缬d
-    int32 action_type=4;//1-点攒，2-取消点攒
-}
-
-message ActionResponse{
-    int32 status_code=1;//状态码，0-成功，其他值失败
-    string status_msg=2;//返▣状态描述
-}
-
-// /douyin/favorite_list - 点赞列表接口
-// 接口类型 GTE
-
-message ListRequest{
-    int64 user_id=1;//用户id
-    string token=2;//用户鉴token
-    }
-message ListResponse{
-    int32 status_code=1;//状态码，O-成功，其他值-失败
-    string status_msg=2;//返▣状态描述
-    Video video_list=3;//用户点赞视颇列表
-    }
-    
-message Video{
-     int64 id=1;//视颇睢一标识
-     User author=2;//视颇/作者信息
-     string play_url=3;//视频播放地址
-     string cover_url=4;//视顷封面地址
-     int64 favorite_count=5;//视频的点赞总数
-     int64 comment_count=6;//视频的评论总数
-     bool is_favorite=7;//true-已点赞，false-末点攒
-     string title=8;//视频标题
-    }
-message User{
-     int64 id=1;//用户id
-     string name=2;//用户名称
-     int64 follow_count=3;//关注总数
-     int64 follower_count=4;//粉丝总数
-     bool is_follow=5;//true-已关注，false-未关注
-    }
-    
-service FavoriteService{
-     rpc Action(ActionRequest)returns (ActionResponse){}
-     rpc List (ListRequest) returns (ListResponse){}
-    }
